legacy: extract RepoObj1, update all imports, drop re-export stubs
