+++ src/borg/helpers/fs.py
@@ -487,7 +487,7 @@ def safe_unlink(path):
     Use this when deleting potentially large files when recovering
     from a VFS error such as ENOSPC. It can help a full file system
     recover. Refer to the "File system interaction" section
-    in legacyrepository.py for further explanations.
+    in legacy/repository.py for further explanations.
     """
     path_obj = Path(path)
     try:
